fast_api_drone/requirements.txt
fastapi==0.95.2
pyyaml==6.0
pymavlink==2.4.0
uvicorn==0.22.0